Separate VirtualProjectBuilder from VirtualProjectBuildingCommand
