+++ src/Cli/dotnet/Commands/Run/RunTelemetry.cs
@@ -6,6 +6,7 @@ using Microsoft.Build.Execution;
 using Microsoft.DotNet.Cli.Utils;
 using Microsoft.DotNet.FileBasedPrograms;
 using Microsoft.DotNet.ProjectTools;
+using Microsoft.DotNet.Utilities;
 
 namespace Microsoft.DotNet.Cli.Commands.Run;
