+++ src/Microsoft.DotNet.ProjectTools/Microsoft.DotNet.ProjectTools.csproj
@@ -12,12 +12,14 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.Build" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" />
   </ItemGroup>
 
   <ItemGroup>
     <InternalsVisibleTo Include="dotnet" />
     <InternalsVisibleTo Include="dotnet.Tests" />
+    <InternalsVisibleTo Include="Microsoft.DotNet.Cli.Utils"/>
     <InternalsVisibleTo Include="Microsoft.DotNet.HotReload.Watch" />
   </ItemGroup>
