+++ test/Microsoft.NET.Build.Tests/Microsoft.NET.Build.Tests.csproj
@@ -35,6 +35,7 @@
 
   <ItemGroup>
     <!-- Make sure tasks project is built, but don't directly reference it as an assembly. -->
+    <ProjectReference Include="..\..\src\Microsoft.DotNet.ProjectTools\Microsoft.DotNet.ProjectTools.csproj" />
     <ProjectReference Include="..\..\src\Tasks\Microsoft.NET.Build.Tasks\Microsoft.NET.Build.Tasks.csproj" PrivateAssets="all" ReferenceOutputAssembly="false" />
 
     <ProjectReference Include="..\Microsoft.NET.TestFramework\Microsoft.NET.TestFramework.csproj" />
